refactor(access-policies): unify CIR Manager and Privilege Map into Access Policies
This commit consolidates the CIR Manager and Privilege Map domains into a unified Access Policies domain. Both originally relied on the identical UserNasPrivilegeMap database model. This consolidation simplifies the architecture by introducing unified routers, services, frontend components, and schemas under 'access_policies', while safely retaining existing data through an Alembic table rename migration.
