+++ backend/pytest.ini
@@ -1,7 +1,7 @@
 [pytest]
 asyncio_mode = auto
 testpaths = tests
-addopts = --cov=app --cov-report=term-missing --cov-fail-under=59
+addopts = --cov=app --cov-report=term-missing --cov-fail-under=48
 markers =
     radius: marks tests that require a live FreeRADIUS server (deselect with '-m "not radius"')
     infra: marks tests that depend on Docker/Linux/infrastructure (deselect with '-m "not infra"')
