+++ backend/requirements.txt
@@ -4,12 +4,12 @@ sqlalchemy==2.0.25
 aiomysql==0.2.0
 alembic==1.13.1
 python-multipart>=0.0.9
-pydantic==2.5.3
-pydantic-settings==2.1.0
+pydantic>=2.5.3
+pydantic-settings>=2.1.0
 cryptography>=42.0.0
 PyJWT>=2.8.0
 bcrypt==4.0.1
-mysqlclient==2.2.1
+mysqlclient>=2.2.1
 pyrad==2.4
 docker==7.1.0
 psutil==5.9.8
