+++ e2e/tests/policies.spec.js
@@ -16,41 +16,71 @@ async function loginAs(page, username, password) {
   }
 }
 
-test.describe('Access Policies & Segments', () => {
+test.describe('Access Policies & Bandwidth Profiles', () => {
   test.beforeEach(async ({ page }) => {
     await loginAs(page, TEST_USER, TEST_PASS)
   })
 
-  test('can create a network segment and exception policy', async ({ page }) => {
+  test('can create a network segment, a bandwidth profile, and an assignment', async ({ page }) => {
     const uniqueSeg = `E2E Seg ${Date.now()}`
+    const profileName = `E2E-100M-${Date.now()}`
 
+    // 1. Create a Network Segment
     await page.goto('/network-segments')
     await expect(page.getByRole('heading', { name: /network segments/i })).toBeVisible()
 
     await page.getByRole('button', { name: /add segment/i }).click()
     await page.getByPlaceholder('e.g. Core Network').fill(uniqueSeg)
-    await page.getByPlaceholder('e.g. 10.0.0.0/8').fill('10.250.0.0/24')
+    await page.getByPlaceholder('e.g. 10.0.0.0/8').fill(`10.${Math.floor(Math.random() * 255)}.0.0/24`)
     await page.getByPlaceholder('Optional description').fill('E2E Testing Segment')
     await page.getByRole('button', { name: /^save$/i }).click()
 
     await expect(page.getByText(uniqueSeg)).toBeVisible({ timeout: 8000 })
 
-    await page.goto('/privilege-map')
+    // 2. Go to Access Policies
+    await page.goto('/access-policies')
     await expect(page.getByRole('heading', { name: /access policies/i })).toBeVisible()
 
+    // 3. Create a Bandwidth Profile
+    await page.getByRole('button', { name: /bandwidth profiles/i }).click()
+    
+    await page.getByLabel('Profile Name').fill(profileName)
+    await page.getByLabel('Downlink High').fill('100M')
+    await page.getByLabel('Uplink High').fill('50M')
+    await page.getByLabel('Downlink Low').fill('10M')
+    await page.getByLabel('Uplink Low').fill('5M')
+    
+    await page.getByRole('button', { name: /save profile/i }).click()
+    await expect(page.getByText(/bandwidth profile saved/i)).toBeVisible({ timeout: 8000 })
+
+    // 4. Create an Assignment
+    await page.getByRole('button', { name: /^assignments$/i }).click()
+    
     await page.getByRole('button', { name: /add policy/i }).click()
     await page.locator('#pm-username').selectOption('jperez')
+    
+    // Target Mode: IP or Range (Exception)
     await page.getByRole('button', { name: /ip or range \(exception\)/i }).click()
-    await page.getByRole('combobox').nth(1).selectOption({ label: `${uniqueSeg} (10.250.0.0/24)` })
+    
+    // Select the segment
+    // It's the select immediately after target mode buttons, but we can find it by its default option
+    await page.getByRole('combobox').filter({ hasText: '-- Select Parent Segment --' }).selectOption({ label: new RegExp(uniqueSeg) })
+    
     await page.getByPlaceholder('e.g. 10.0.1.50').fill('10.250.0.10')
-    await page.locator('#pm-radius-group').selectOption({ index: 1 })
+    
+    // Select the newly created Bandwidth Profile
+    await page.locator('#pm-radius-group').selectOption({ label: profileName })
+    
     await page.locator('#pm-approved-by').fill('E2E Test')
+    
     await page.getByRole('button', { name: /save policy/i }).click()
 
     await expect(page.getByRole('heading', { name: /new policy/i })).toBeHidden({ timeout: 8000 })
 
+    // Verify assignment exists in the list for the user
     await page.getByRole('button', { name: /jperez/i }).click()
     await expect(page.getByText(uniqueSeg)).toBeVisible({ timeout: 8000 })
     await expect(page.getByText('10.250.0.10')).toBeVisible()
+    await expect(page.getByText(profileName)).toBeVisible()
   })
 })
