+++ frontend/src/components/Layout.jsx
@@ -23,8 +23,7 @@ const Layout = () => {
         { name: 'NAS Devices', path: '/nas', icon: Server },
         { name: 'Active Sessions', path: '/sessions', icon: Activity },
         { name: 'Groups', path: '/groups', icon: Layers },
-        { name: 'Access Policies', path: '/privilege-map', icon: Map, allowedRoles: ['superadmin', 'admin', 'auditor'] },
-        { name: 'CIR Manager', path: '/cir', icon: Gauge, allowedRoles: ['superadmin', 'admin', 'auditor'] },
+        { name: 'Access Policies', path: '/access-policies', icon: Map, allowedRoles: ['superadmin', 'admin', 'auditor'] },
         { name: 'Network Segments', path: '/network-segments', icon: Server, allowedRoles: ['superadmin', 'admin'] },
         { name: 'IAM & Zonas', path: '/iam', icon: Shield, allowedRoles: ['superadmin', 'admin'] },
         { name: 'Dictionaries', path: '/dictionaries', icon: BookOpen },
